+++ src/components/Admin/AddProduct.vue
@@ -3,23 +3,13 @@
     <!-- Area Chart -->
     <div class="col-xl-12 col-lg-12">
       <div class="card shadow mb-4">
-        <!-- Card Header - Dropdown -->
+       
         <div
           class="card-header py-3 d-flex flex-row align-items-center justify-content-between"
-        >
-          <h6 class="m-0 font-weight-bold text-primary">
-            You are about to ADD New Product
-          </h6>
-        </div>
+        ></div>
         <!-- Card Body -->
         <div class="card-body">
-          <div class="adding-table">
-            <AddTable />
-          </div>
-          <div class="overview-table">
-            <OverviewTable 
-            />
-          </div>
+          <AddProductItem />
         </div>
       </div>
     </div>
@@ -27,13 +17,11 @@
 </template>
 
 <script>
-import AddTable from "@/components/Admin/AddTable.vue";
-import OverviewTable from "@/components/Admin/OverviewTable";
+import AddProductItem from "@/components/Admin/AddProductItem.vue";
+
 export default {
-    
   components: {
-    AddTable,
-    OverviewTable,
+    AddProductItem,
   },
 };
 </script>
